imgproc: C-API cleanup pt.4
