+++ modules/imgproc/include/opencv2/imgproc/imgproc_c.h
@@ -53,157 +53,11 @@ extern "C" {
 @{
 */
 
-/****************************************************************************************\
-*                                    Image Processing                                    *
-\****************************************************************************************/
-
-/*********************************** data sampling **************************************/
-
-/****************************************************************************************\
-*                              Contours retrieving                                       *
-\****************************************************************************************/
-
-/** @brief Retrieves outer and optionally inner boundaries of white (non-zero) connected
-   components in the black (zero) background
-@see cv::findContours, cvStartFindContours, cvFindNextContour, cvSubstituteContour, cvEndFindContours
-*/
-CVAPI(int)  cvFindContours( CvArr* image, CvMemStorage* storage, CvSeq** first_contour,
-                            int header_size CV_DEFAULT(sizeof(CvContour)),
-                            int mode CV_DEFAULT(CV_RETR_LIST),
-                            int method CV_DEFAULT(CV_CHAIN_APPROX_SIMPLE),
-                            CvPoint offset CV_DEFAULT(cvPoint(0,0)));
-
-/** @brief Initializes contour retrieving process.
-
-   Calls cvStartFindContours.
-   Calls cvFindNextContour until null pointer is returned
-   or some other condition becomes true.
-   Calls cvEndFindContours at the end.
-@see cvFindContours
-*/
-CVAPI(CvContourScanner)  cvStartFindContours( CvArr* image, CvMemStorage* storage,
-                            int header_size CV_DEFAULT(sizeof(CvContour)),
-                            int mode CV_DEFAULT(CV_RETR_LIST),
-                            int method CV_DEFAULT(CV_CHAIN_APPROX_SIMPLE),
-                            CvPoint offset CV_DEFAULT(cvPoint(0,0)));
-
-/** @brief Retrieves next contour
-@see cvFindContours
-*/
-CVAPI(CvSeq*)  cvFindNextContour( CvContourScanner scanner );
-
-
-/** @brief Substitutes the last retrieved contour with the new one
-
-   (if the substitutor is null, the last retrieved contour is removed from the tree)
-@see cvFindContours
-*/
-CVAPI(void)   cvSubstituteContour( CvContourScanner scanner, CvSeq* new_contour );
-
-
-/** @brief Releases contour scanner and returns pointer to the first outer contour
-@see cvFindContours
-*/
-CVAPI(CvSeq*)  cvEndFindContours( CvContourScanner* scanner );
-
-/** @brief Approximates Freeman chain(s) with a polygonal curve.
-
-This is a standalone contour approximation routine, not represented in the new interface. When
-cvFindContours retrieves contours as Freeman chains, it calls the function to get approximated
-contours, represented as polygons.
-
-@param src_seq Pointer to the approximated Freeman chain that can refer to other chains.
-@param storage Storage location for the resulting polylines.
-@param method Approximation method (see the description of the function :ocvFindContours ).
-@param parameter Method parameter (not used now).
-@param minimal_perimeter Approximates only those contours whose perimeters are not less than
-minimal_perimeter . Other chains are removed from the resulting structure.
-@param recursive Recursion flag. If it is non-zero, the function approximates all chains that can
-be obtained from chain by using the h_next or v_next links. Otherwise, the single input chain is
-approximated.
-@see cvStartReadChainPoints, cvReadChainPoint
- */
-CVAPI(CvSeq*) cvApproxChains( CvSeq* src_seq, CvMemStorage* storage,
-                            int method CV_DEFAULT(CV_CHAIN_APPROX_SIMPLE),
-                            double parameter CV_DEFAULT(0),
-                            int  minimal_perimeter CV_DEFAULT(0),
-                            int  recursive CV_DEFAULT(0));
-
-/** @brief Initializes Freeman chain reader.
-
-   The reader is used to iteratively get coordinates of all the chain points.
-   If the Freeman codes should be read as is, a simple sequence reader should be used
-@see cvApproxChains
-*/
-CVAPI(void) cvStartReadChainPoints( CvChain* chain, CvChainPtReader* reader );
-
-/** @brief Retrieves the next chain point
-@see cvApproxChains
-*/
-CVAPI(CvPoint) cvReadChainPoint( CvChainPtReader* reader );
 
 
 /****************************************************************************************\
 *                            Contour Processing and Shape Analysis                       *
 \****************************************************************************************/
-
-/** @brief Approximates a single polygonal curve (contour) or
-   a tree of polygonal curves (contours)
-@see cv::approxPolyDP
-*/
-CVAPI(CvSeq*)  cvApproxPoly( const void* src_seq,
-                             int header_size, CvMemStorage* storage,
-                             int method, double eps,
-                             int recursive CV_DEFAULT(0));
-
-/** @brief Calculates perimeter of a contour or length of a part of contour
-@see cv::arcLength
-*/
-CVAPI(double)  cvArcLength( const void* curve,
-                            CvSlice slice CV_DEFAULT(CV_WHOLE_SEQ),
-                            int is_closed CV_DEFAULT(-1));
-
-/** same as cvArcLength for closed contour
-*/
-CV_INLINE double cvContourPerimeter( const void* contour )
-{
-    return cvArcLength( contour, CV_WHOLE_SEQ, 1 );
-}
-
-
-/** @brief Calculates contour bounding rectangle (update=1) or
-   just retrieves pre-calculated rectangle (update=0)
-@see cv::boundingRect
-*/
-CVAPI(CvRect)  cvBoundingRect( CvArr* points, int update CV_DEFAULT(0) );
-
-/** @brief Calculates area of a contour or contour segment
-@see cv::contourArea
-*/
-CVAPI(double)  cvContourArea( const CvArr* contour,
-                              CvSlice slice CV_DEFAULT(CV_WHOLE_SEQ),
-                              int oriented CV_DEFAULT(0));
-
-/** @brief Calculates exact convex hull of 2d point set
-@see cv::convexHull
-*/
-CVAPI(CvSeq*) cvConvexHull2( const CvArr* input,
-                             void* hull_storage CV_DEFAULT(NULL),
-                             int orientation CV_DEFAULT(CV_CLOCKWISE),
-                             int return_points CV_DEFAULT(0));
-
-/** @brief Checks whether the contour is convex or not (returns 1 if convex, 0 if not)
-@see cv::isContourConvex
-*/
-CVAPI(int)  cvCheckContourConvexity( const CvArr* contour );
-
-
-/** @brief Initializes sequence header for a matrix (column or row vector) of points
-
-   a wrapper for cvMakeSeqHeaderForArray (it does not initialize bounding rectangle!!!) */
-CVAPI(CvSeq*) cvPointSeqFromMat( int seq_kind, const CvArr* mat,
-                                 CvContour* contour_header,
-                                 CvSeqBlock* block );
 
 /****************************************************************************************\
 *                                  Histogram functions                                   *
