+++ modules/imgproc/src/approx.cpp
@@ -53,6 +53,28 @@ typedef struct _CvPtInfo
 }
 _CvPtInfo;
 
+static const CvPoint icvCodeDeltas[8] =
+    { {1, 0}, {1, -1}, {0, -1}, {-1, -1}, {-1, 0}, {-1, 1}, {0, 1}, {1, 1} };
+
+static void cvStartReadChainPoints( CvChain * chain, CvChainPtReader * reader )
+{
+    int i;
+
+    if( !chain || !reader )
+        CV_Error( CV_StsNullPtr, "" );
+
+    if( chain->elem_size != 1 || chain->header_size < (int)sizeof(CvChain))
+        CV_Error( CV_StsBadSize, "" );
+
+    cvStartReadSeq( (CvSeq *) chain, (CvSeqReader *) reader, 0 );
+
+    reader->pt = chain->origin;
+    for( i = 0; i < 8; i++ )
+    {
+        reader->deltas[i][0] = (schar) icvCodeDeltas[i].x;
+        reader->deltas[i][1] = (schar) icvCodeDeltas[i].y;
+    }
+}
 
 /* curvature: 0 - 1-curvature, 1 - k-cosine curvature. */
 CvSeq* icvApproximateChainTC89( CvChain* chain, int header_size,
@@ -374,96 +396,6 @@ copy_vect:
     return cvEndWriteSeq( &writer );
 }
 
-
-/*Applies some approximation algorithm to chain-coded contour(s) and
-  converts it/them to polygonal representation */
-CV_IMPL CvSeq*
-cvApproxChains( CvSeq*              src_seq,
-                CvMemStorage*       storage,
-                int                 method,
-                double              /*parameter*/,
-                int                 minimal_perimeter,
-                int                 recursive )
-{
-    CvSeq *prev_contour = 0, *parent = 0;
-    CvSeq *dst_seq = 0;
-
-    if( !src_seq || !storage )
-        CV_Error( CV_StsNullPtr, "" );
-    if( method > CV_CHAIN_APPROX_TC89_KCOS || method <= 0 || minimal_perimeter < 0 )
-        CV_Error( CV_StsOutOfRange, "" );
-
-    while( src_seq != 0 )
-    {
-        int len = src_seq->total;
-
-        if( len >= minimal_perimeter )
-        {
-            CvSeq *contour = 0;
-
-            switch( method )
-            {
-            case CV_CHAIN_APPROX_NONE:
-            case CV_CHAIN_APPROX_SIMPLE:
-            case CV_CHAIN_APPROX_TC89_L1:
-            case CV_CHAIN_APPROX_TC89_KCOS:
-                contour = icvApproximateChainTC89( (CvChain *) src_seq, sizeof( CvContour ), storage, method );
-                break;
-            default:
-                CV_Error( CV_StsOutOfRange, "" );
-            }
-
-            if( contour->total > 0 )
-            {
-                cvBoundingRect( contour, 1 );
-
-                contour->v_prev = parent;
-                contour->h_prev = prev_contour;
-
-                if( prev_contour )
-                    prev_contour->h_next = contour;
-                else if( parent )
-                    parent->v_next = contour;
-                prev_contour = contour;
-                if( !dst_seq )
-                    dst_seq = prev_contour;
-            }
-            else                /* if resultant contour has zero length, skip it */
-            {
-                len = -1;
-            }
-        }
-
-        if( !recursive )
-            break;
-
-        if( src_seq->v_next && len >= minimal_perimeter )
-        {
-            CV_Assert( prev_contour != 0 );
-            parent = prev_contour;
-            prev_contour = 0;
-            src_seq = src_seq->v_next;
-        }
-        else
-        {
-            while( src_seq->h_next == 0 )
-            {
-                src_seq = src_seq->v_prev;
-                if( src_seq == 0 )
-                    break;
-                prev_contour = parent;
-                if( parent )
-                    parent = parent->v_prev;
-            }
-            if( src_seq )
-                src_seq = src_seq->h_next;
-        }
-    }
-
-    return dst_seq;
-}
-
-
 /****************************************************************************************\
 *                               Polygonal Approximation                                  *
 \****************************************************************************************/
@@ -708,156 +640,3 @@ void cv::approxPolyDP( InputArray _curve, OutputArray _approxCurve,
 
     Mat(nout, 1, CV_MAKETYPE(depth, 2), buf).copyTo(_approxCurve);
 }
-
-
-CV_IMPL CvSeq*
-cvApproxPoly( const void* array, int header_size,
-             CvMemStorage* storage, int method,
-             double parameter, int parameter2 )
-{
-    cv::AutoBuffer<cv::Point> _buf;
-    cv::AutoBuffer<cv::Range> stack(100);
-    CvSeq* dst_seq = 0;
-    CvSeq *prev_contour = 0, *parent = 0;
-    CvContour contour_header;
-    CvSeq* src_seq = 0;
-    CvSeqBlock block;
-    int recursive = 0;
-
-    if( CV_IS_SEQ( array ))
-    {
-        src_seq = (CvSeq*)array;
-        if( !CV_IS_SEQ_POLYLINE( src_seq ))
-            CV_Error( CV_StsBadArg, "Unsupported sequence type" );
-
-        recursive = parameter2;
-
-        if( !storage )
-            storage = src_seq->storage;
-    }
-    else
-    {
-        src_seq = cvPointSeqFromMat(
-                                    CV_SEQ_KIND_CURVE | (parameter2 ? CV_SEQ_FLAG_CLOSED : 0),
-                                    array, &contour_header, &block );
-    }
-
-    if( !storage )
-        CV_Error( CV_StsNullPtr, "NULL storage pointer " );
-
-    if( header_size < 0 )
-        CV_Error( CV_StsOutOfRange, "header_size is negative. "
-                 "Pass 0 to make the destination header_size == input header_size" );
-
-    if( header_size == 0 )
-        header_size = src_seq->header_size;
-
-    if( !CV_IS_SEQ_POLYLINE( src_seq ))
-    {
-        if( CV_IS_SEQ_CHAIN( src_seq ))
-        {
-            CV_Error( CV_StsBadArg, "Input curves are not polygonal. "
-                     "Use cvApproxChains first" );
-        }
-        else
-        {
-            CV_Error( CV_StsBadArg, "Input curves have unknown type" );
-        }
-    }
-
-    if( header_size == 0 )
-        header_size = src_seq->header_size;
-
-    if( header_size < (int)sizeof(CvContour) )
-        CV_Error( CV_StsBadSize, "New header size must be non-less than sizeof(CvContour)" );
-
-    if( method != CV_POLY_APPROX_DP )
-        CV_Error( CV_StsOutOfRange, "Unknown approximation method" );
-
-    while( src_seq != 0 )
-    {
-        CvSeq *contour = 0;
-
-        switch (method)
-        {
-        case CV_POLY_APPROX_DP:
-            if( parameter < 0 )
-                CV_Error( CV_StsOutOfRange, "Accuracy must be non-negative" );
-
-            CV_Assert( CV_SEQ_ELTYPE(src_seq) == CV_32SC2 ||
-                      CV_SEQ_ELTYPE(src_seq) == CV_32FC2 );
-
-            {
-            int npoints = src_seq->total, nout = 0;
-            _buf.allocate(npoints*2);
-            cv::Point *src = _buf.data(), *dst = src + npoints;
-            bool closed = CV_IS_SEQ_CLOSED(src_seq);
-
-            if( src_seq->first->next == src_seq->first )
-                src = (cv::Point*)src_seq->first->data;
-            else
-                cvCvtSeqToArray(src_seq, src);
-
-            if( CV_SEQ_ELTYPE(src_seq) == CV_32SC2 )
-                nout = cv::approxPolyDP_(src, npoints, dst, closed, parameter, stack);
-            else if( CV_SEQ_ELTYPE(src_seq) == CV_32FC2 )
-                nout = cv::approxPolyDP_((cv::Point2f*)src, npoints,
-                                         (cv::Point2f*)dst, closed, parameter, stack);
-            else
-                CV_Error( CV_StsUnsupportedFormat, "" );
-
-            contour = cvCreateSeq( src_seq->flags, header_size,
-                                  src_seq->elem_size, storage );
-            cvSeqPushMulti(contour, dst, nout);
-            }
-            break;
-        default:
-            CV_Error( CV_StsBadArg, "Invalid approximation method" );
-        }
-
-        CV_Assert( contour );
-
-        if( header_size >= (int)sizeof(CvContour))
-            cvBoundingRect( contour, 1 );
-
-        contour->v_prev = parent;
-        contour->h_prev = prev_contour;
-
-        if( prev_contour )
-            prev_contour->h_next = contour;
-        else if( parent )
-            parent->v_next = contour;
-        prev_contour = contour;
-        if( !dst_seq )
-            dst_seq = prev_contour;
-
-        if( !recursive )
-            break;
-
-        if( src_seq->v_next )
-        {
-            CV_Assert( prev_contour != 0 );
-            parent = prev_contour;
-            prev_contour = 0;
-            src_seq = src_seq->v_next;
-        }
-        else
-        {
-            while( src_seq->h_next == 0 )
-            {
-                src_seq = src_seq->v_prev;
-                if( src_seq == 0 )
-                    break;
-                prev_contour = parent;
-                if( parent )
-                    parent = parent->v_prev;
-            }
-            if( src_seq )
-                src_seq = src_seq->h_next;
-        }
-    }
-
-    return dst_seq;
-}
-
-/* End of file. */
