+++ modules/imgproc/src/convhull.cpp
@@ -455,140 +455,3 @@ bool isContourConvex( InputArray _contour )
 }
 
 }
-
-CV_IMPL CvSeq*
-cvConvexHull2( const CvArr* array, void* hull_storage,
-               int orientation, int return_points )
-{
-    CvMat* mat = 0;
-    CvContour contour_header;
-    CvSeq hull_header;
-    CvSeqBlock block, hullblock;
-    CvSeq* ptseq = 0;
-    CvSeq* hullseq = 0;
-
-    if( CV_IS_SEQ( array ))
-    {
-        ptseq = (CvSeq*)array;
-        if( !CV_IS_SEQ_POINT_SET( ptseq ))
-            CV_Error( CV_StsBadArg, "Unsupported sequence type" );
-        if( hull_storage == 0 )
-            hull_storage = ptseq->storage;
-    }
-    else
-    {
-        ptseq = cvPointSeqFromMat( CV_SEQ_KIND_GENERIC, array, &contour_header, &block );
-    }
-
-    bool isStorage = isStorageOrMat(hull_storage);
-
-    if(isStorage)
-    {
-        if( return_points )
-        {
-            hullseq = cvCreateSeq(CV_SEQ_KIND_CURVE|CV_SEQ_ELTYPE(ptseq)|
-                                  CV_SEQ_FLAG_CLOSED|CV_SEQ_FLAG_CONVEX,
-                                  sizeof(CvContour), sizeof(CvPoint),(CvMemStorage*)hull_storage );
-        }
-        else
-        {
-            hullseq = cvCreateSeq(
-                                  CV_SEQ_KIND_CURVE|CV_SEQ_ELTYPE_PPOINT|
-                                  CV_SEQ_FLAG_CLOSED|CV_SEQ_FLAG_CONVEX,
-                                  sizeof(CvContour), sizeof(CvPoint*), (CvMemStorage*)hull_storage );
-        }
-    }
-    else
-    {
-        mat = (CvMat*)hull_storage;
-
-        if( (mat->cols != 1 && mat->rows != 1) || !CV_IS_MAT_CONT(mat->type))
-            CV_Error( CV_StsBadArg,
-                     "The hull matrix should be continuous and have a single row or a single column" );
-
-        if( mat->cols + mat->rows - 1 < ptseq->total )
-            CV_Error( CV_StsBadSize, "The hull matrix size might be not enough to fit the hull" );
-
-        if( CV_MAT_TYPE(mat->type) != CV_SEQ_ELTYPE(ptseq) &&
-           CV_MAT_TYPE(mat->type) != CV_32SC1 )
-            CV_Error( CV_StsUnsupportedFormat,
-                     "The hull matrix must have the same type as input or 32sC1 (integers)" );
-
-        hullseq = cvMakeSeqHeaderForArray(
-                                          CV_SEQ_KIND_CURVE|CV_MAT_TYPE(mat->type)|CV_SEQ_FLAG_CLOSED,
-                                          sizeof(hull_header), CV_ELEM_SIZE(mat->type), mat->data.ptr,
-                                          mat->cols + mat->rows - 1, &hull_header, &hullblock );
-        cvClearSeq( hullseq );
-    }
-
-    int hulltype = CV_SEQ_ELTYPE(hullseq);
-    int total = ptseq->total;
-    if( total == 0 )
-    {
-        if( !isStorage )
-            CV_Error( CV_StsBadSize,
-                     "Point sequence can not be empty if the output is matrix" );
-        return 0;
-    }
-
-    cv::AutoBuffer<double> _ptbuf;
-    cv::Mat h0;
-    cv::convexHull(cv::cvarrToMat(ptseq, false, false, 0, &_ptbuf), h0,
-                   orientation == CV_CLOCKWISE, CV_MAT_CN(hulltype) == 2);
-
-
-    if( hulltype == CV_SEQ_ELTYPE_PPOINT )
-    {
-        const int* idx = h0.ptr<int>();
-        int ctotal = (int)h0.total();
-        for( int i = 0; i < ctotal; i++ )
-        {
-            void* ptr = cvGetSeqElem(ptseq, idx[i]);
-            cvSeqPush( hullseq, &ptr );
-        }
-    }
-    else
-        cvSeqPushMulti(hullseq, h0.ptr(), (int)h0.total());
-
-    if (isStorage)
-    {
-        return hullseq;
-    }
-    else
-    {
-        if( mat->rows > mat->cols )
-            mat->rows = hullseq->total;
-        else
-            mat->cols = hullseq->total;
-        return 0;
-    }
-}
-
-
-CV_IMPL int
-cvCheckContourConvexity( const CvArr* array )
-{
-    CvContour contour_header;
-    CvSeqBlock block;
-    CvSeq* contour = (CvSeq*)array;
-
-    if( CV_IS_SEQ(contour) )
-    {
-        if( !CV_IS_SEQ_POINT_SET(contour))
-            CV_Error( CV_StsUnsupportedFormat,
-                     "Input sequence must be polygon (closed 2d curve)" );
-    }
-    else
-    {
-        contour = cvPointSeqFromMat(CV_SEQ_KIND_CURVE|
-            CV_SEQ_FLAG_CLOSED, array, &contour_header, &block );
-    }
-
-    if( contour->total == 0 )
-        return -1;
-
-    cv::AutoBuffer<double> _buf;
-    return cv::isContourConvex(cv::cvarrToMat(contour, false, false, 0, &_buf)) ? 1 : 0;
-}
-
-/* End of file. */
