+++ modules/imgproc/src/shapedescr.cpp
@@ -861,659 +861,3 @@ cv::RotatedRect cv::fitEllipseDirect( InputArray _points )
     }
     return box;
 }
-
-
-namespace cv
-{
-
-// Calculates bounding rectangle of a point set or retrieves already calculated
-static Rect pointSetBoundingRect( const Mat& points )
-{
-    int npoints = points.checkVector(2);
-    int depth = points.depth();
-    CV_Assert(npoints >= 0 && (depth == CV_32F || depth == CV_32S));
-
-    int  xmin = 0, ymin = 0, xmax = -1, ymax = -1, i;
-    bool is_float = depth == CV_32F;
-
-    if( npoints == 0 )
-        return Rect();
-
-#if CV_SIMD
-    const int64_t* pts = points.ptr<int64_t>();
-
-    if( !is_float )
-    {
-        v_int32 minval, maxval;
-        minval = maxval = v_reinterpret_as_s32(vx_setall_s64(*pts)); //min[0]=pt.x, min[1]=pt.y, min[2]=pt.x, min[3]=pt.y
-        for( i = 1; i <= npoints - v_int32::nlanes/2; i+= v_int32::nlanes/2 )
-        {
-            v_int32 ptXY2 = v_reinterpret_as_s32(vx_load(pts + i));
-            minval = v_min(ptXY2, minval);
-            maxval = v_max(ptXY2, maxval);
-        }
-        minval = v_min(v_reinterpret_as_s32(v_expand_low(v_reinterpret_as_u32(minval))), v_reinterpret_as_s32(v_expand_high(v_reinterpret_as_u32(minval))));
-        maxval = v_max(v_reinterpret_as_s32(v_expand_low(v_reinterpret_as_u32(maxval))), v_reinterpret_as_s32(v_expand_high(v_reinterpret_as_u32(maxval))));
-        if( i <= npoints - v_int32::nlanes/4 )
-        {
-            v_int32 ptXY = v_reinterpret_as_s32(v_expand_low(v_reinterpret_as_u32(vx_load_low(pts + i))));
-            minval = v_min(ptXY, minval);
-            maxval = v_max(ptXY, maxval);
-            i += v_int64::nlanes/2;
-        }
-        for(int j = 16; j < CV_SIMD_WIDTH; j*=2)
-        {
-            minval = v_min(v_reinterpret_as_s32(v_expand_low(v_reinterpret_as_u32(minval))), v_reinterpret_as_s32(v_expand_high(v_reinterpret_as_u32(minval))));
-            maxval = v_max(v_reinterpret_as_s32(v_expand_low(v_reinterpret_as_u32(maxval))), v_reinterpret_as_s32(v_expand_high(v_reinterpret_as_u32(maxval))));
-        }
-        xmin = minval.get0();
-        xmax = maxval.get0();
-        ymin = v_reinterpret_as_s32(v_expand_high(v_reinterpret_as_u32(minval))).get0();
-        ymax = v_reinterpret_as_s32(v_expand_high(v_reinterpret_as_u32(maxval))).get0();
-#if CV_SIMD_WIDTH > 16
-        if( i < npoints )
-        {
-            v_int32x4 minval2, maxval2;
-            minval2 = maxval2 = v_reinterpret_as_s32(v_expand_low(v_reinterpret_as_u32(v_load_low(pts + i))));
-            for( i++; i < npoints; i++ )
-            {
-                v_int32x4 ptXY = v_reinterpret_as_s32(v_expand_low(v_reinterpret_as_u32(v_load_low(pts + i))));
-                minval2 = v_min(ptXY, minval2);
-                maxval2 = v_max(ptXY, maxval2);
-            }
-            xmin = min(xmin, minval2.get0());
-            xmax = max(xmax, maxval2.get0());
-            ymin = min(ymin, v_reinterpret_as_s32(v_expand_high(v_reinterpret_as_u32(minval2))).get0());
-            ymax = max(ymax, v_reinterpret_as_s32(v_expand_high(v_reinterpret_as_u32(maxval2))).get0());
-        }
-#endif
-    }
-    else
-    {
-        v_float32 minval, maxval;
-        minval = maxval = v_reinterpret_as_f32(vx_setall_s64(*pts)); //min[0]=pt.x, min[1]=pt.y, min[2]=pt.x, min[3]=pt.y
-        for( i = 1; i <= npoints - v_float32::nlanes/2; i+= v_float32::nlanes/2 )
-        {
-            v_float32 ptXY2 = v_reinterpret_as_f32(vx_load(pts + i));
-            minval = v_min(ptXY2, minval);
-            maxval = v_max(ptXY2, maxval);
-        }
-        minval = v_min(v_reinterpret_as_f32(v_expand_low(v_reinterpret_as_u32(minval))), v_reinterpret_as_f32(v_expand_high(v_reinterpret_as_u32(minval))));
-        maxval = v_max(v_reinterpret_as_f32(v_expand_low(v_reinterpret_as_u32(maxval))), v_reinterpret_as_f32(v_expand_high(v_reinterpret_as_u32(maxval))));
-        if( i <= npoints - v_float32::nlanes/4 )
-        {
-            v_float32 ptXY = v_reinterpret_as_f32(v_expand_low(v_reinterpret_as_u32(vx_load_low(pts + i))));
-            minval = v_min(ptXY, minval);
-            maxval = v_max(ptXY, maxval);
-            i += v_float32::nlanes/4;
-        }
-        for(int j = 16; j < CV_SIMD_WIDTH; j*=2)
-        {
-            minval = v_min(v_reinterpret_as_f32(v_expand_low(v_reinterpret_as_u32(minval))), v_reinterpret_as_f32(v_expand_high(v_reinterpret_as_u32(minval))));
-            maxval = v_max(v_reinterpret_as_f32(v_expand_low(v_reinterpret_as_u32(maxval))), v_reinterpret_as_f32(v_expand_high(v_reinterpret_as_u32(maxval))));
-        }
-        xmin = cvFloor(minval.get0());
-        xmax = cvFloor(maxval.get0());
-        ymin = cvFloor(v_reinterpret_as_f32(v_expand_high(v_reinterpret_as_u32(minval))).get0());
-        ymax = cvFloor(v_reinterpret_as_f32(v_expand_high(v_reinterpret_as_u32(maxval))).get0());
-#if CV_SIMD_WIDTH > 16
-        if( i < npoints )
-        {
-            v_float32x4 minval2, maxval2;
-            minval2 = maxval2 = v_reinterpret_as_f32(v_expand_low(v_reinterpret_as_u32(v_load_low(pts + i))));
-            for( i++; i < npoints; i++ )
-            {
-                v_float32x4 ptXY = v_reinterpret_as_f32(v_expand_low(v_reinterpret_as_u32(v_load_low(pts + i))));
-                minval2 = v_min(ptXY, minval2);
-                maxval2 = v_max(ptXY, maxval2);
-            }
-            xmin = min(xmin, cvFloor(minval2.get0()));
-            xmax = max(xmax, cvFloor(maxval2.get0()));
-            ymin = min(ymin, cvFloor(v_reinterpret_as_f32(v_expand_high(v_reinterpret_as_u32(minval2))).get0()));
-            ymax = max(ymax, cvFloor(v_reinterpret_as_f32(v_expand_high(v_reinterpret_as_u32(maxval2))).get0()));
-        }
-#endif
-    }
-#else
-    const Point* pts = points.ptr<Point>();
-    Point pt = pts[0];
-
-    if( !is_float )
-    {
-        xmin = xmax = pt.x;
-        ymin = ymax = pt.y;
-
-        for( i = 1; i < npoints; i++ )
-        {
-            pt = pts[i];
-
-            if( xmin > pt.x )
-                xmin = pt.x;
-
-            if( xmax < pt.x )
-                xmax = pt.x;
-
-            if( ymin > pt.y )
-                ymin = pt.y;
-
-            if( ymax < pt.y )
-                ymax = pt.y;
-        }
-    }
-    else
-    {
-        Cv32suf v;
-        // init values
-        xmin = xmax = CV_TOGGLE_FLT(pt.x);
-        ymin = ymax = CV_TOGGLE_FLT(pt.y);
-
-        for( i = 1; i < npoints; i++ )
-        {
-            pt = pts[i];
-            pt.x = CV_TOGGLE_FLT(pt.x);
-            pt.y = CV_TOGGLE_FLT(pt.y);
-
-            if( xmin > pt.x )
-                xmin = pt.x;
-
-            if( xmax < pt.x )
-                xmax = pt.x;
-
-            if( ymin > pt.y )
-                ymin = pt.y;
-
-            if( ymax < pt.y )
-                ymax = pt.y;
-        }
-
-        v.i = CV_TOGGLE_FLT(xmin); xmin = cvFloor(v.f);
-        v.i = CV_TOGGLE_FLT(ymin); ymin = cvFloor(v.f);
-        // because right and bottom sides of the bounding rectangle are not inclusive
-        // (note +1 in width and height calculation below), cvFloor is used here instead of cvCeil
-        v.i = CV_TOGGLE_FLT(xmax); xmax = cvFloor(v.f);
-        v.i = CV_TOGGLE_FLT(ymax); ymax = cvFloor(v.f);
-    }
-#endif
-
-    return Rect(xmin, ymin, xmax - xmin + 1, ymax - ymin + 1);
-}
-
-
-static Rect maskBoundingRect( const Mat& img )
-{
-    CV_Assert( img.depth() <= CV_8S && img.channels() == 1 );
-
-    Size size = img.size();
-    int xmin = size.width, ymin = -1, xmax = -1, ymax = -1, i, j, k;
-
-    for( i = 0; i < size.height; i++ )
-    {
-        const uchar* _ptr = img.ptr(i);
-        const uchar* ptr = (const uchar*)alignPtr(_ptr, 4);
-        int have_nz = 0, k_min, offset = (int)(ptr - _ptr);
-        j = 0;
-        offset = MIN(offset, size.width);
-        for( ; j < offset; j++ )
-            if( _ptr[j] )
-            {
-                have_nz = 1;
-                break;
-            }
-        if( j < offset )
-        {
-            if( j < xmin )
-                xmin = j;
-            if( j > xmax )
-                xmax = j;
-        }
-        if( offset < size.width )
-        {
-            xmin -= offset;
-            xmax -= offset;
-            size.width -= offset;
-            j = 0;
-            for( ; j <= xmin - 4; j += 4 )
-                if( *((int*)(ptr+j)) )
-                    break;
-            for( ; j < xmin; j++ )
-                if( ptr[j] )
-                {
-                    xmin = j;
-                    if( j > xmax )
-                        xmax = j;
-                    have_nz = 1;
-                    break;
-                }
-            k_min = MAX(j-1, xmax);
-            k = size.width - 1;
-            for( ; k > k_min && (k&3) != 3; k-- )
-                if( ptr[k] )
-                    break;
-            if( k > k_min && (k&3) == 3 )
-            {
-                for( ; k > k_min+3; k -= 4 )
-                    if( *((int*)(ptr+k-3)) )
-                        break;
-            }
-            for( ; k > k_min; k-- )
-                if( ptr[k] )
-                {
-                    xmax = k;
-                    have_nz = 1;
-                    break;
-                }
-            if( !have_nz )
-            {
-                j &= ~3;
-                for( ; j <= k - 3; j += 4 )
-                    if( *((int*)(ptr+j)) )
-                        break;
-                for( ; j <= k; j++ )
-                    if( ptr[j] )
-                    {
-                        have_nz = 1;
-                        break;
-                    }
-            }
-            xmin += offset;
-            xmax += offset;
-            size.width += offset;
-        }
-        if( have_nz )
-        {
-            if( ymin < 0 )
-                ymin = i;
-            ymax = i;
-        }
-    }
-
-    if( xmin >= size.width )
-        xmin = ymin = 0;
-    return Rect(xmin, ymin, xmax - xmin + 1, ymax - ymin + 1);
-}
-
-}
-
-cv::Rect cv::boundingRect(InputArray array)
-{
-    CV_INSTRUMENT_REGION();
-
-    Mat m = array.getMat();
-    return m.depth() <= CV_8U ? maskBoundingRect(m) : pointSetBoundingRect(m);
-}
-
-////////////////////////////////////////////// C API ///////////////////////////////////////////
-
-static void
-icvMemCopy( double **buf1, double **buf2, double **buf3, int *b_max )
-{
-    CV_Assert( (*buf1 != NULL || *buf2 != NULL) && *buf3 != NULL );
-
-    int bb = *b_max;
-    if( *buf2 == NULL )
-    {
-        *b_max = 2 * (*b_max);
-        *buf2 = (double *)cvAlloc( (*b_max) * sizeof( double ));
-
-        memcpy( *buf2, *buf3, bb * sizeof( double ));
-
-        *buf3 = *buf2;
-        cvFree( buf1 );
-        *buf1 = NULL;
-    }
-    else
-    {
-        *b_max = 2 * (*b_max);
-        *buf1 = (double *) cvAlloc( (*b_max) * sizeof( double ));
-
-        memcpy( *buf1, *buf3, bb * sizeof( double ));
-
-        *buf3 = *buf1;
-        cvFree( buf2 );
-        *buf2 = NULL;
-    }
-}
-
-
-/* area of a contour sector */
-static double icvContourSecArea( CvSeq * contour, CvSlice slice )
-{
-    cv::Point pt;                 /*  pointer to points   */
-    cv::Point pt_s, pt_e;         /*  first and last points  */
-    CvSeqReader reader;         /*  points reader of contour   */
-
-    int p_max = 2, p_ind;
-    int lpt, flag, i;
-    double a00;                 /* unnormalized moments m00    */
-    double xi, yi, xi_1, yi_1, x0, y0, dxy, sk, sk1, t;
-    double x_s, y_s, nx, ny, dx, dy, du, dv;
-    double eps = 1.e-5;
-    double *p_are1, *p_are2, *p_are;
-    double area = 0;
-
-    CV_Assert( contour != NULL && CV_IS_SEQ_POINT_SET( contour ));
-
-    lpt = cvSliceLength( slice, contour );
-    /*if( n2 >= n1 )
-        lpt = n2 - n1 + 1;
-    else
-        lpt = contour->total - n1 + n2 + 1;*/
-
-    if( contour->total <= 0 || lpt <= 2 )
-        return 0.;
-
-    a00 = x0 = y0 = xi_1 = yi_1 = 0;
-    sk1 = 0;
-    flag = 0;
-    dxy = 0;
-    p_are1 = (double *) cvAlloc( p_max * sizeof( double ));
-
-    p_are = p_are1;
-    p_are2 = NULL;
-
-    cvStartReadSeq( contour, &reader, 0 );
-    cvSetSeqReaderPos( &reader, slice.start_index );
-    { CvPoint pt_s_ = CV_STRUCT_INITIALIZER; CV_READ_SEQ_ELEM(pt_s_, reader); pt_s = pt_s_; }
-    p_ind = 0;
-    cvSetSeqReaderPos( &reader, slice.end_index );
-    { CvPoint pt_e_ = CV_STRUCT_INITIALIZER; CV_READ_SEQ_ELEM(pt_e_, reader); pt_e = pt_e_; }
-
-/*    normal coefficients    */
-    nx = pt_s.y - pt_e.y;
-    ny = pt_e.x - pt_s.x;
-    cvSetSeqReaderPos( &reader, slice.start_index );
-
-    while( lpt-- > 0 )
-    {
-        { CvPoint pt_ = CV_STRUCT_INITIALIZER; CV_READ_SEQ_ELEM(pt_, reader); pt = pt_; }
-
-        if( flag == 0 )
-        {
-            xi_1 = (double) pt.x;
-            yi_1 = (double) pt.y;
-            x0 = xi_1;
-            y0 = yi_1;
-            sk1 = 0;
-            flag = 1;
-        }
-        else
-        {
-            xi = (double) pt.x;
-            yi = (double) pt.y;
-
-/****************   edges intersection examination   **************************/
-            sk = nx * (xi - pt_s.x) + ny * (yi - pt_s.y);
-            if( (fabs( sk ) < eps && lpt > 0) || sk * sk1 < -eps )
-            {
-                if( fabs( sk ) < eps )
-                {
-                    dxy = xi_1 * yi - xi * yi_1;
-                    a00 = a00 + dxy;
-                    dxy = xi * y0 - x0 * yi;
-                    a00 = a00 + dxy;
-
-                    if( p_ind >= p_max )
-                        icvMemCopy( &p_are1, &p_are2, &p_are, &p_max );
-
-                    p_are[p_ind] = a00 / 2.;
-                    p_ind++;
-                    a00 = 0;
-                    sk1 = 0;
-                    x0 = xi;
-                    y0 = yi;
-                    dxy = 0;
-                }
-                else
-                {
-/*  define intersection point    */
-                    dv = yi - yi_1;
-                    du = xi - xi_1;
-                    dx = ny;
-                    dy = -nx;
-                    if( fabs( du ) > eps )
-                        t = ((yi_1 - pt_s.y) * du + dv * (pt_s.x - xi_1)) /
-                            (du * dy - dx * dv);
-                    else
-                        t = (xi_1 - pt_s.x) / dx;
-                    if( t > eps && t < 1 - eps )
-                    {
-                        x_s = pt_s.x + t * dx;
-                        y_s = pt_s.y + t * dy;
-                        dxy = xi_1 * y_s - x_s * yi_1;
-                        a00 += dxy;
-                        dxy = x_s * y0 - x0 * y_s;
-                        a00 += dxy;
-                        if( p_ind >= p_max )
-                            icvMemCopy( &p_are1, &p_are2, &p_are, &p_max );
-
-                        p_are[p_ind] = a00 / 2.;
-                        p_ind++;
-
-                        a00 = 0;
-                        sk1 = 0;
-                        x0 = x_s;
-                        y0 = y_s;
-                        dxy = x_s * yi - xi * y_s;
-                    }
-                }
-            }
-            else
-                dxy = xi_1 * yi - xi * yi_1;
-
-            a00 += dxy;
-            xi_1 = xi;
-            yi_1 = yi;
-            sk1 = sk;
-
-        }
-    }
-
-    xi = x0;
-    yi = y0;
-    dxy = xi_1 * yi - xi * yi_1;
-
-    a00 += dxy;
-
-    if( p_ind >= p_max )
-        icvMemCopy( &p_are1, &p_are2, &p_are, &p_max );
-
-    p_are[p_ind] = a00 / 2.;
-    p_ind++;
-
-    // common area calculation
-    area = 0;
-    for( i = 0; i < p_ind; i++ )
-        area += fabs( p_are[i] );
-
-    if( p_are1 != NULL )
-        cvFree( &p_are1 );
-    else if( p_are2 != NULL )
-        cvFree( &p_are2 );
-
-    return area;
-}
-
-
-/* external contour area function */
-CV_IMPL double
-cvContourArea( const void *array, CvSlice slice, int oriented )
-{
-    double area = 0;
-
-    CvContour contour_header;
-    CvSeq* contour = 0;
-    CvSeqBlock block;
-
-    if( CV_IS_SEQ( array ))
-    {
-        contour = (CvSeq*)array;
-        if( !CV_IS_SEQ_POLYLINE( contour ))
-            CV_Error( CV_StsBadArg, "Unsupported sequence type" );
-    }
-    else
-    {
-        contour = cvPointSeqFromMat( CV_SEQ_KIND_CURVE, array, &contour_header, &block );
-    }
-
-    if( cvSliceLength( slice, contour ) == contour->total )
-    {
-        cv::AutoBuffer<double> abuf;
-        cv::Mat points = cv::cvarrToMat(contour, false, false, 0, &abuf);
-        return cv::contourArea( points, oriented !=0 );
-    }
-
-    if( CV_SEQ_ELTYPE( contour ) != CV_32SC2 )
-        CV_Error( CV_StsUnsupportedFormat,
-        "Only curves with integer coordinates are supported in case of contour slice" );
-    area = icvContourSecArea( contour, slice );
-    return oriented ? area : fabs(area);
-}
-
-
-/* calculates length of a curve (e.g. contour perimeter) */
-CV_IMPL  double
-cvArcLength( const void *array, CvSlice slice, int is_closed )
-{
-    double perimeter = 0;
-
-    int i, j = 0, count;
-    const int N = 16;
-    float buf[N];
-    CvMat buffer = cvMat( 1, N, CV_32F, buf );
-    CvSeqReader reader;
-    CvContour contour_header;
-    CvSeq* contour = 0;
-    CvSeqBlock block;
-
-    if( CV_IS_SEQ( array ))
-    {
-        contour = (CvSeq*)array;
-        if( !CV_IS_SEQ_POLYLINE( contour ))
-            CV_Error( CV_StsBadArg, "Unsupported sequence type" );
-        if( is_closed < 0 )
-            is_closed = CV_IS_SEQ_CLOSED( contour );
-    }
-    else
-    {
-        is_closed = is_closed > 0;
-        contour = cvPointSeqFromMat(
-                                    CV_SEQ_KIND_CURVE | (is_closed ? CV_SEQ_FLAG_CLOSED : 0),
-                                    array, &contour_header, &block );
-    }
-
-    if( contour->total > 1 )
-    {
-        int is_float = CV_SEQ_ELTYPE( contour ) == CV_32FC2;
-
-        cvStartReadSeq( contour, &reader, 0 );
-        cvSetSeqReaderPos( &reader, slice.start_index );
-        count = cvSliceLength( slice, contour );
-
-        count -= !is_closed && count == contour->total;
-
-        // scroll the reader by 1 point
-        reader.prev_elem = reader.ptr;
-        CV_NEXT_SEQ_ELEM( sizeof(CvPoint), reader );
-
-        for( i = 0; i < count; i++ )
-        {
-            float dx, dy;
-
-            if( !is_float )
-            {
-                CvPoint* pt = (CvPoint*)reader.ptr;
-                CvPoint* prev_pt = (CvPoint*)reader.prev_elem;
-
-                dx = (float)pt->x - (float)prev_pt->x;
-                dy = (float)pt->y - (float)prev_pt->y;
-            }
-            else
-            {
-                CvPoint2D32f* pt = (CvPoint2D32f*)reader.ptr;
-                CvPoint2D32f* prev_pt = (CvPoint2D32f*)reader.prev_elem;
-
-                dx = pt->x - prev_pt->x;
-                dy = pt->y - prev_pt->y;
-            }
-
-            reader.prev_elem = reader.ptr;
-            CV_NEXT_SEQ_ELEM( contour->elem_size, reader );
-            // Bugfix by Axel at rubico.com 2010-03-22, affects closed slices only
-            // wraparound not handled by CV_NEXT_SEQ_ELEM
-            if( is_closed && i == count - 2 )
-                cvSetSeqReaderPos( &reader, slice.start_index );
-
-            buffer.data.fl[j] = dx * dx + dy * dy;
-            if( ++j == N || i == count - 1 )
-            {
-                buffer.cols = j;
-                cvPow( &buffer, &buffer, 0.5 );
-                for( ; j > 0; j-- )
-                    perimeter += buffer.data.fl[j-1];
-            }
-        }
-    }
-
-    return perimeter;
-}
-
-
-/* Calculates bounding rectangle of a point set or retrieves already calculated */
-CV_IMPL  CvRect
-cvBoundingRect( CvArr* array, int update )
-{
-    cv::Rect rect;
-    CvContour contour_header;
-    CvSeq* ptseq = 0;
-    CvSeqBlock block;
-
-    CvMat stub, *mat = 0;
-    int calculate = update;
-
-    if( CV_IS_SEQ( array ))
-    {
-        ptseq = (CvSeq*)array;
-        if( !CV_IS_SEQ_POINT_SET( ptseq ))
-            CV_Error( CV_StsBadArg, "Unsupported sequence type" );
-
-        if( ptseq->header_size < (int)sizeof(CvContour))
-        {
-            update = 0;
-            calculate = 1;
-        }
-    }
-    else
-    {
-        mat = cvGetMat( array, &stub );
-        if( CV_MAT_TYPE(mat->type) == CV_32SC2 ||
-            CV_MAT_TYPE(mat->type) == CV_32FC2 )
-        {
-            ptseq = cvPointSeqFromMat(CV_SEQ_KIND_GENERIC, mat, &contour_header, &block);
-            mat = 0;
-        }
-        else if( CV_MAT_TYPE(mat->type) != CV_8UC1 &&
-                CV_MAT_TYPE(mat->type) != CV_8SC1 )
-            CV_Error( CV_StsUnsupportedFormat,
-                "The image/matrix format is not supported by the function" );
-        update = 0;
-        calculate = 1;
-    }
-
-    if( !calculate )
-        return ((CvContour*)ptseq)->rect;
-
-    if( mat )
-    {
-        rect = cvRect(cv::maskBoundingRect(cv::cvarrToMat(mat)));
-    }
-    else if( ptseq->total )
-    {
-        cv::AutoBuffer<double> abuf;
-        rect = cvRect(cv::pointSetBoundingRect(cv::cvarrToMat(ptseq, false, false, 0, &abuf)));
-    }
-    if( update )
-        ((CvContour*)ptseq)->rect = cvRect(rect);
-    return cvRect(rect);
-}
-
-/* End of file. */
